[debugger] Work on async formatting.
There should be no behavior change.

Adds a completion callback to the FillFormatNodeDescription. This is
currently always synchronous but for pointers and strings this will need
to be asynchronous in the future. Adding this callback will allow those
implementations.

Uses fit::callback for this new use and the other uses in the new formatting
system (format.* and format_node.*). Since the completion callback has no
arguments, we can use the deferred_action helper to always issue the callback
on completion which eliminates the need to manually call the callback or track
it through execution.

The call sites using deferred_action are a little nasty, I proposed a
patch to fit in a separate patch to make the constructor implicit that
will hopefully clean these up.

Converts the legacy formatting system to use the async version of
FillFormatNodeDescription. These are temporarily a bit nasty. When the
new system is complete, there will be two separate passes: the first an
async pass to fill the tree, then the final outputting pass will be
simpler synchronous code.

Since the new code that uses fit::callback needs to call older code that
uses std::callback in a few places, a helper function is added to
convert. I plan on converting all single-call callbacks to
fit::function which will eventually make this conversion function
obsolete.

Change-Id: Ib7094eb546a32994c3e101ef51c8047fa4e14103