[bootserver] -1 means run once to completion
Currently there are cases where the protocol will become wedged due to races
between restarts, or lost packets. When this occurs, using the -1 flag is
prohibitive, as it exits too early. The documented intent, and the common
user intent is to run until one completion, rather than only one attempt.

The existing behavior is useful for development of paving itself, but is not
put behind the --fail-fast flag.

Change-Id: I93fd558dd9e588c6f7a942cbb13fc6f017d8e4ab