[build] Make package() from other toolchains redirect, not pass deps through
If a package() target is a dependency of some target in a toolchain other
than $target_toolchain, the old behavior was to turn it into a group()
target with deps copied from the invoker.  This leads to deps in the other
toolchain that were never intended to be instantiated anywhere but
package().  It also fails to provide a dependency on the actual packaging
actions.  Instead, it should redirect to the same package() target in the
$target_toolchain so that its deps are evaluated in the expected context.

Bug: BLD-534 #done
Change-Id: I6216d9acc2ed2e48e3b7a48d1cc44566c770fd44