[codec] Quit() codec runner when done
Without this, we basically leaked the runner for SW-based codecs.

So far I haven't thought of a way to test this change, so it may sit
un-merged until we find a way to test it.  Maybe if we had an optional
additional channel or eventpair routed to the same server process that's
guaranteed to only close on process exit...

Change-Id: I0c08355465aa3a4278ea66ff04ffde9ae358293c