[banjo][syzkaller] Add support for syzkaller pointers
This CL adds support for strings, arrays and multiple returns in Banjo.
These are all translated to pointers in Syzkaller descriptions.

Removed support for floating points. The approach from this point on is
to only support features required by sycalls.banjo. Following this,
arrays and strings are not supported as additional return types.

Common code between Syzkaller and Abigen backend was moved to util.rs
file.

Syzkaller only outputs return types, if it is used as input argument for
some other syscall (called resource [0]). Backend was updated to not
output any return types. Future support for resources will update this.

[0]: https://github.com/google/syzkaller/blob/master/docs/syscall_descriptions_syntax.md#resources

SEC-326 #done

Test: fx run-host-tests banjo_unittests

Change-Id: Id47e84d0946dd547d9b5bc4ab3b55f7bf155b661