[zircon][core-test] Fix flake timeout.
Because this test is not expecting a response, there is a race condition
between the calling thread timing out, and the replying thread being
destructed. Because the replying thread owns the handle, if it gets
destructed before the channel times out, it will get PEER_CLOSED instead
of ERR_TIMED_OUT.

Added an alternative ReplyAndWait which stalls the replying thread,
until the calling thread signals an event, which prevents the handle
from being closed.

// Manual Validation:
25k iters of the test without failures.

FLK-394 #done

TEST=channel-test

Change-Id: I9ec135e93c959adbe557afdd8b32d1a4e8a07e97