[componentmgr] Implement Hub as Capability
This CL permits a component to request a view of the Hub rooted at it.

The follow syntax in a component's cml file requests this view:

"use": [
  {
    "directory": "/hub/exec", "from": "framework", "as": "/hub"
  }
]

This notation says take the exec directory of this component's hub
and map it to /hub in this component's namespace.

What's really powerful here is a component does not need to map the whole
hub rooted at it but can map ANY subdirectory.

Change-Id: I70487a45f0f05aaf5440d1887c44c0d76faaee58