[identity] AccountManager supports ephemeral
This enables support for ephemeral accounts in the AccountManager,
i.e. accounts that are scoped to the duration of a session, or at the
very most lives until the end of a power cycle. Account data for an
ephemeral account is stored in memory only. The opposite of an ephemeral
account is a persistent account.

Changes:
- AccountManager now opens connection to a the (new)
  `account_handler_ephemeral` component for any ephemeral accounts.
- For now, `lifetime` is stored as an immutable property of the
  `AccountHandlerConnection`.
- AccountManager uses this property to filter out persistent accounts
  when it writes its account list to disk. The in-memory map
  (`ids_to_handlers`) contains both ephemeral and persistent accounts.

AUTH-76 #comment

Change-Id: Ib35790eeed2ee67d078cb649817f877763a8e926