[zircon][core-test]: Fix flaky ChannelInternalTest
There is a race condition in a channel call when the channel is written
to and the thread state is set to blocked. Instead of just asserting
for the thread state, we wait for the channel to become readable, and
then wait for the thread state to be updated.
Essentially the expectation the thread is blocked switches to the thread
will eventually become blocked when issuing a channel_call.

ZX-4422 #done

TEST=channel-test

Change-Id: Iffa110f67a63cc6ba450631ceafa7430aa44f885