[build] Export shlib-compatible builds of Zircon static libraries
The static libraries built in Zircon and then exported for use in
the legacy build can sometimes be linked into shared libraries
there.  Make sure to export the shlib-compatible versions of
those libraries, so those links are valid.  Within Zircon, the
correct shlib or non-shlib version of a static library is always
found as needed because of the toolchain setup.  But that gets
flattened to a single binary for each library in the legacy
export.  The shlib-compatible build is always fine (if not
optimal) to use in non-shlib contexts, but not vice versa.

Bug: BLD-537 #done
Change-Id: I3801703eb5f501001e093a6d72d7123fe828bd91