[virtualization] Temporarily disable virtio input tests on ARM.
The ARM linux image doesn't expose "/dev/input/inputX" devices,
preventing the in-guest test scripts from detecting keystrokes.

Until we work out what is going on, disable the tests on ARM64.

In the meantime, we still have some virtio-input test coverage in the
GPU tests (which ensure that keystrokes causes the console to change).

Bug: MAC-256, ZX-4414
Change-Id: I8d57271d31230d88d973e7e1d610b83331c53a12