[async-testing] abstract subloop interface
This change enables running any kind of subloop inside TestLoop, as long
as it provides the right operations. Operations are provided through a
vtable-like C structure so they can be implemented through another
language's FFI.
The interface between the test loop and the subloops remains almost
identical, except that the timekeeper is removed: instead, the testloop
notifies all subloops when time advances.

Test: existing implementation tested by existing tests, new test for
using an external loop

Change-Id: Ie10d115a5d4a04a9f67a53973bcb5fe9b36681ff