[scenic] add and use pessimistic frame scheduler
This change replaces the optimistic frame scheduler with a pessimistic
one, which means we use the largest time out of the last N frames as the
next prediction, instead of the smallest time. We also begin using the
frame scheduler instead of the 8ms hardcoded offset.

SCN-1415 #comment
SCN-960 #done

Test: peridot+topaz input latency benchmarks, manual tracing, unit tests

Change-Id: I1b412487863c4af93dc93ef3f3f9431391a148b6