[media][drm] Add decryptor params
In order for the fuchsia.media.StreamProcessor to be capable of usage
for decryption, it needs some additional data about the format of
encrypted data. For example, it needs to know which encryption mode and
key to use. This CL adds these additional metadata items the
fuchsia.media.FormatDetails object.

MTWN-322 #comment

Change-Id: I74a91ebd91b062ba10da7a07de6a1bb560d51aa3