[netdump] Make filter tests generic
The existing packet filter tests are made generic so that they
are decoupled from how filters are constructed. The motivation is
that filters constructed using different methods but with the same
parameters should have the same behavior, and should pass the same
battery of tests. The tests are extracted as templates in a header
file, with a filter construction C++ callable that must be supplied.
Repeated instantiations of the same filter tests can then be made
for different approaches in constructing the filters, with minimal
repetion of code.

This CL includes instantiations of the filter tests for filter tree
node constructors, thereby keeping observable test behavior the same
as before. In a follow-up CL, the same tests will be instantiated
for filters constructed with the filter builder interface.

Bug: CONN-99
Test: Build with `--with-base //garnet/packages/tests:zircon` as a
parameter to `fx set`, then `runtests -t netdump-test` on the target.

Change-Id: Ice796ac9e5d9be6583071a029aace537ea0db608