[netdump] Concrete filter builder implementation
This CL implements the `FilterBuilder` interface concretely for
filter tree nodes. That interface is used by the parser to build
a filter during the parse of a filter string.

Testing of the CL is done through instantiating the generic set
of filter tests exposed by fxr/295509. The test exposed a bug in
`tokens.h` where the tag value of the `HOST` token was incorrect,
and this was fixed.

Bug: CONN-99
Test: Build with `--with-base //garnet/packages/tests:zircon` as a
parameter to `fx set`, then `runtests -t netdump-test` on the target.

Change-Id: I4a578863532fd4d0cbdae07ec177da288e9b81ee