[netstack] Implement fuchsia.net.NameLookup
This implements the fuchsia.net.NameLookup protocol in the Go netstack.
The implementation of LookupIp is complete, while the implementation of
LookupHostname is only correct for localhost right now, since the
netstack DNS client implementation does not currently support reverse DNS
lookups.

This commit was submitted as a4fb557, but caused network problems, and
was reverted in 761ee17. The original commit added locking around
tryOneName, causing deadlock and flakiness. This commit removes that
locking, restoring the code to its original behaviour.

Change-Id: Ic1d660bd19865f3a280b8468dbcd4fbd6b62c4f0