[kernel] Handle sparse GIC v2 maps.
On ARM platforms interrupts are dispatched by the GIC. v2 of this ARM
standard does not specify how cpus are assigned to the cpu_mask inside
of the GIC and thus how they are to be addressed.

For most platforms we operate on, our logical cpu id (and mask) inside of
the kernel happens to match the one inside of the GIC. This was not the
case on sherlock however.

We will capture this data in the topology tree passed from the
bootloader, the field existed there but was not plumbed through. We will
then initialize a local map as the cpus come up and sue this map to
translate requests sent to the GIC on GIC v2 platforms.

There is a newer GIC v3 that defines how these masks are assigned so we
don't need to do this on those arm platforms. This is why we keep the
logic constrained to the v2 implementation.

Test:
* Added kernel unit test for new logic.
* Booted on sherlock, with proper bootloader change, and verified that
all cores come up, system works as expected and "k ut all" passes.
* Booted on Astro, with current bootloader, and verified that all cores
still come up, system works as expected, and "k ut all" passes.
Issue:
ZX-4236 #System support

Change-Id: I4c311c7f1caca978e79330f059a9c90d4788a2fb