[brcmfmac] Rename debugging macros to upper-case
This is purely a formatting change to bring the driver up to style
guide recommendations that macro names should be all-uppercase.  Also
in this change is a rename of BRCMF_FOO_ON() to BRCMF_IS_ON(FOO), in
preparation for a later refactor of debug.h.

TESTED=built for vim2

Change-Id: I79fb7812d870dc13e424d3de734f3cbe57823647