[debugger] Correct stepping multiple threads over a breakpoint.
This implemented correct stepping(tm). This is meant to address the
problem that GDB's displaced stepping solves: when a thread is stepping
over a breakpoint, the latter is removed and another thread could pass
through it. This is actually easily replicated by the
"MultithreadedBreakpoint" integration test (not reenabled in this CL).

The strategy here is that when a thread begins stepping over, all other
threads will be suspended, except those that are currently stepping over
a breakpoint. The moment a thread is done stepping over a breakpoint, it
is suspended until all threads are done stepping over the breakpoint.
Once all threads are over the breakpoint, it is reinstalled and the
threads resumed.

Due to waiting for threads to be suspended, the debugger can be sure
that all threads are suspended and no race should affect this. The only
edge case is that there is a race between suspending a thread and it
starting another thread, which won't necessarilly be tracked by the
breakpoint and could theoretically pass through it. That is quite an
edge case and is not addressed by this CL.

Added a long unit test verifying this behaviour.

DX-1500  #comment
DX-1065  #comment

TEST=Added unittest.
     fx run-test debug_agent_tests.

Change-Id: Ia33626bd75aedd198e02da9273732736cc8947b7