[netclock] Initiate a roughtime sync when starting.
network_time is currently an app started in the network state
by sysmgr. We'd like to delete this and rely exclusively on
network_time_service for roughtime updates, but to do so we need
to make sure that *someone* is starting a time update on boot.

fuchsia.time.Utc is a startup service implemented by netclock,
causing netclock to start early in the boot flow. On netclock's
start, we'll attempt to fetch roughtime updates in a loop with
exponential backoff. This in turn launches network_time_service,
obviating the need for a boot-specific binary to be launched.

This change is an important step to making netclock the actual
system time authority, and it makes fuchsia.time.Utc useful.

CF-835 #done
CF-850 #done

Change-Id: Iec28d8e15a7331de2061b3607f4e44018c818dc3