[netdump] Netdump filter syntax description
In this CL, the packet filter language syntax is described in two
locations, both for the benefit of human-readability when the output
is rendered:

First, it is described in a `SYNTAX.md` file that will be later
absorbed into a `README.md` document.

Second, writing the syntax to an output stream is implemented in a
`parser_syntax` function designed to be called when the user wishes
to see the filter usage. Currently, this function is called in a test
and the output examined for correctness. This test will be removed
once the use of this function is integrated into netdump.

Bug: CONN-99, CONN-145
Test: Build with `--with-base //garnet/packages/tests:zircon` as a
parameter to `fx set`, then `runtests -t netdump-test` on the target.
The printed output was manually examined.

Change-Id: I8d1cacd57b78088cc0ba32099b225b5cc01c80e4