[audio] Add Audio Usages
This adds per usage gain attributes and includes them in all streams.
In addition, it extends the AudioCore fidl interface to control them.

Usages can be changed at any time on both capturers and renderers, but
care should be used when changing the usage of a stream that is active,
because the new usage may be applied to audio data that is currently
buffered.

Test: fx build, pave
      fx run-test audio_fidl_tests
      fx run-test audio_mixer_tests
      fx run-test mediaplayer_tests
Change-Id: Ibfe357f076193e0ddf18a587841ec4a334694b0a