[magma] Set min_coded_width/height to 0
The input sizes are actually the minimum necessary sizes the buffer is
expected to require, not the minimum sizes the driver supports (which
are assumed to be 0). This allows setting varying the size given to
magma_get_buffer_format_plane_info_with_size, but means that
magma_get_buffer_format_plane_info will always return stride 0 and
should be removed.

Change-Id: Ia341285c8e44ededa889359b841430e4e9827c19