[virtualization] Disable buffering of stdout.
By default, stdout is buffered by the C library to avoid frequent
write() syscalls. This presents problems in the context of automated
testing, however: if a test hangs, we lose up to ~20 lines of log output
(stuck in a buffer), making it hard to determine where precisely the
test failed.

We've recently seen hangs in the guest virtualization tests, so update
stdout to be line-buffered to help diagnose the issue.

Ideally, we would update stdout buffering globally, instead of just for
this test-suite: IN-1446 tracks a global fix to this issue.

Change-Id: I65969526bb1b332495aa97e0474c7f4933bd9e8b