[LSC][fuchsia-async] maintain executor-specific time
The global executor now maintains its own notion of time: it is either
the actual system time (for execution) or a fake time (useful for
testing). The executor-relative time can be queried indepently of
whether it is fake using fasync::Time::now(). If the executor is in
fake-time mode, its internal time is advanced manually.
To avoid mixing up executor-relative time with the time used in system
calls, the executor-relative time is made incompatible with zx::Time,
and all timers require an executor-relative time. This requires updating
all uses of zx::Duration::after_now to either import fasync::DurationExt
or use zx::Time::after.

Change-Id: I1101abb6cee3f810da1e89d552d542647cdbed6f