[sherlock][audio] Fix pop on start of audio play
ZX-4549 #done

Test: The behavior change is the removal of a pop heard through a
speaker when audio starts, until we have an automated way of tapping
into the output of the codec, we need to manually check that no pop
is heard.  Used uapp audio tone with a sine wave (audio gain -65 and
audio -c 4 tone) via netboot in a sherlock with bringup config and
heard no pop.  Also heard no pop while booting on a paved system.

Change-Id: I8254fe92115453be08791db025938353182f762b