[netstack3] Properly calculate original packet body size for ICMP responses
For IPv6/ICMPv6 responses, RFC 4443 requires us to send the original packet
back in the ICMP response's message body without the packet exceeding the
IPv6 minimum MTU.

Before this change, the size of the ICMP message body was allowed to be up
to the IPv6 minimum MTU size. This could cause the resulting IPv6 packet
to have a size greater than the IPv6 minimum MTU because of the IPv6
headers and ICMPv6 headers.

This change passes to the IP packet's body serialization request the IP
header size so it can be used to generate a packet body without exceeding
some size given an IP header's size.

Tests: Validating that when attempting to forward a packet of size greater
than the device MTU, a Packet Too Big message is sent as a reply with the
original packet in the ICMP message body after being shrunk to fit within
the IPv6 minimum MTU.

Change-Id: I30ddbc19b63464c309ea5d5dfd226a2bc360e636