[zircon][serial] correct uart16550 locking and state notifications
Notifications are now provided for readability and writability up and
down states. Prior to this change, notifications were only given for the
up state for each. This was in conflict with the expectations of the
primary serial driver.

This change also corrects for logical race conditions resulting from the
new state handling by using a single mutex with broader critical
sections.

Change-Id: I7d5acdd6c5b6623141ed03962fcf9fea980fa5b0