[zircon] Make netboot vmos resizable
When netbooting, the command line is appended to the data vmo. If the
existing size + the command line length happens to change the number of
pages required for the combined VMO, then netbooting would fail because
the VMO wasn't resizable.

This might be only one cause of netboot failures, as there are
un-bug-documented reports of this happening before work on ZX-2357
started, but this locally fixes a repro I had on an arm device.

Adds new test to libzbi/test/zbi.cc, but it could use more.
ZbiVMO::SplitComplete() is otherwise untested before this CL.

ZX-2357 #comment [zircon] Make netboot vmos resizable
ZX-4597 #done

Change-Id: Ia3215cda6e2f409081f5f26e20b9e6d306f45f0f