[pkgfs] Add "ctl" subdir
The current API for triggering a GC of pkgfs is to unlink
/pkgfs/garbage, but this API is incompatible with mapping
"/pkgfs/garbage" into a component's sandbox, since unlinking a file is
an operation on the parent directory, which cannot be mapped into a
component's sandbox without depending on the deprecated shell feature.

This change moves /pkgfs/garbage and /pkgfs/validation under a common
directory: /pkgfs/ctl. For backwards compatibility with existing users,
/pkgfs/garbage and /pkgfs/validation will continue to exist until those
users are migrated to the new location.

Bug: PKG-804 #comment add ctl dir
Change-Id: I9df86e9641f08e3b7187b476c9e6bbca770f1933