[build] Add zbi_test() template, use it for core-tests
Generalize building a standalone ZBI test a la core-tests.  The
template bakes in the kernel command line arguments and does some
metadata niceties that integrate with the legacy build and its
existing infra protocols to get ZBI tests run on bots.

userboot.reboot is changed to take precedence over userboot.shutdown
so that userboot.shutdown can be baked into the image (which is
ideal for QEMU) but still be overriden at boot time if e.g. booting
the ZBI test on hardware where rebooting back to its zero state
rather than powering it off is probably preferable.

When in either reboot or shutdown mode, userboot now prints the exit
status of the launched process and also prints a unique string that
will be matched in the console log to indicate success of the test.

Bug: BLD-382 #comment zbi_tests.json
Change-Id: If3cf582d2602e9cfd761238f4aac969b40b9e5a6