[SetUI] Switchboard and Registry Introduction.
Switchboard is an interface facilitating source driven messaging and
notifications. It is responsible for properly routing requests based
on setting type. Switchboard abstracts the communication substrate,
allowing for future changes.

Registry complements Switchboard by providing an interface for
entities to register as handlers of various setting types. It works
in conjunction with Switchboard to route requests to the proper
handler.

The interaction between Registry and Switchboard create a single
communication channel for all requests and events. This will allow
for future transformations/processing to be added to the pipeline.

Fixes: SU-331
Test: fx run-test setui_service_tests
Change-Id: Ifba097e28806654913a6bafd2d3f231a7d6b7927