[build][rust] Make third-party lib locations knowable at gen time
This change was first landed in d8c047c and reverted in d3c47c9 because
it broke the mac build, due to checking for whether the current target
being built for was mac rather than `host_os == "mac"`. This has been
fixed, and this change should therefore be safe to re-land.

Test: built, ran on mac builders and locally on mac
Change-Id: Idb7c418b2ae45d3a58676d7cf92a05cb4ac4a9d3