[bt][cli] Only show peer events when there is a change in connected/bonded state
Previously, we printed the values for any peer for which we received an
added or updated notification. This resulted in a lot of noisy update
whilst using bt-cli, most of which are irrelevant.

Instead, filter to only those events which show a change in 'connected'
or 'bonded' state, which are the main updates that we care about. This
increases the signal to noise ratio of updates significantly, and
improves the usability of the bt-cli tool.

Change-Id: I13b7af9b64fe40c5774fdfa5476cc6ad368303c2