[fidlc] Types have names
This both simplifies the code, and provides the proper resolved name of
a type. In the case of builtins, we resolve `library/name` to `name`,
i.e. a name in the root library. Previously, it was not easy to access
the properly resolved name.

All types are created by type templates (or should once fully
canonicalized), and all type templates own the name of the type which
they produce. This makes it possible for types simply have references,
and keep ownership with type templates.

FIDL-447 #done

Test: no behavior change
Change-Id: If43b7e1e98ee0c82964536045a0883049f8f0d86