[debugger] Add promises to debugger message loop
The debugger uses a different message loop because it needs to support
compiling on the host. But this meant we did not get fit::promise
support.

This implements fit::executor and associated machinery on the
MessageLoop so promises can be posted. It additionally allows promises
to be run initially bypassing the message queue. This allows us to keep
our current pattern of running such code immediately and only going to
the message loop if it completes asynchronously.

Change-Id: Ifc5b4222374492b4b93fdd0f2d91802158a98352