[hid] Remove hid size test
ZX-4445 #done

The purpose of this test originally was to check
that our full HID parser and our partial HID parser
outputted the same sizes on the same inputs.

At this point, we only have one HID parser, and the
sizes for these devices are verified in the hid-parser
library. This test is a dupe.

TEST: runtests -t hid-parser-test
Change-Id: Iacab09428e96f0a0cd92ecfa6f355caf079c3d6e