[devcoord] Add kSuspending state to device.
This will allow the unbind task to know a device is suspending
and wait for it to finish (to be added in a future CL).
Also allow a device to be removed after it is suspended.

The test helper function CheckSuspendReceived has been split into
a sending and receiving part, which will allow us to test
unbind / suspend being scheduled on the same device simultaneously.

TEST= ran devcoordinator-test

Change-Id: I76b8d6c4fc4139d6222f72318ab89e39919bbae4