[debugger] Fix incorrect resume over breakpoint.
The problem was due that the breakpoing was told step over was done
before resuming from an exception. Due to the way the code is layed out,
this would make the thread attempt to step over it again.

This by itself is not catastrophic, but it would mark the thread as
currently stepping over a breakpoint when it really wasn't. This would
trigger an assert when attempting to step over a *second* breakpoint.
The fix was a one liner.

Manual zxdb testing would not uncover this as normally users step over
code, which clears the state. But fidlcat effectivelly exercises many
breakpoints in a row, uncovering this issue. I was able to replicate
this behaviour manually in the debugger too.

Added more logging that helped identify the issue.
Added a --debug-mode flag to show the DebugAgent's log into stdout when
running fidlcat (really useful for debugging this issue!).

TEST=Extended the breakpoint integration test (that actually inserts
     breakpoints in memory) to exercise this use case. The test does assert
     when the fix is removed and passes with it.

     Run: fx run-test debug_agent_tests

DX-1717 #done

Change-Id: I3f3b1a0fceeb9e5a31b13bcf52789746565c4a21