[graphics][compute]: Fix debug shader file extension.
Changes the extension of debug shaders generated during
the build of the graphics compute project, from .spv to
.glsl.

Note that these are human-readable versions of the optimized
shaders generated by the hotsort and spinel libraries. They
are only used for human review and are not used to build
or run anything.

BUG: SPN-47
TEST: fx build src/graphics/lib/compute:tests &&
      ls out/default/gen/src/graphics/lib/compute/hotsort/platforms/vk/targets/hs_amd_gcn3_u32/debug/*.glsl
Change-Id: I55658451e75dc1bf3ad2a5acda4385b4eeb77b4e