[futex][ownership] Fix an issue which was causing test flake.
Fix an issue in the futex ownership logic.  If a thread attempts to
block on a futex, and assign ownership in the process, but the thread
never actually blocks on the underlying wait queue (because it has
already timed out, or has been signaled to be killed), the thread
_still_ needs to assign ownership of the futex.

Because of the absolute timeouts used by zircon in all syscalls, it is
possible in a very heavily loaded test environment to compute a
timeout, but have the timeout be already expired by the time that the
thread makes it into the syscall (exposing the bug).

++ Fix the bug
++ Rework the timing and some of the loging in tests to help to avoid
   this situation.
++ Add a test which explicitly exercises this edge case.

FLK-427  #done

Change-Id: Ie9ed03f9d8457a52cb3d934eed00a2cafa05ef2b