[wlan-hw-sim] Ensure wlan-if devices are spawned
Since wlancfg will always try to create an iface device when it detects
a phy device, tests intended for the client role should always wait
until wlancfg finishes creating such iface so that
|clear_iface_and_ensure_iface_gone()| can always find the iface and
remove it.

This allows |test_set_country()| to be enabled

Note: This is not added as part of
|test_utils::TestHelper::begin_tests()| because this function is also
used for AP tests which completely bypass wlancfg.

Test: Test-only change that does not change behavior

FLK-416 #done

Change-Id: Ife3bc3c28780abc12f9b95b6d6e6dbf5ce57971a