[kerne][rsrc] Allow shared resources to create new, smaller resources
This change allows a given shared resource to create a smaller resource
of the same kind as long as the base / len of the new resource are
entirely contained within the access range provided by the parent
resource. No parent or child hierarchy is maintained after creation, and
at this time the parent retains access to the region handed out to the
child resource. This serves to enable bus drivers with access to a
resource window to create smaller resources to hand out to their
downstream devices.

To facilitate this change some small reworks and streamlining has been
done with the resource validation methods available in the kernel.

Change-Id: I5820edf0261e1f16ae9ccf25eba52bc0ed909d32