[brcmfmac] Modularize driver components
Thsi commit works to modularize the driver build.  The purpose here is
to reduce the number of places where conditional compilation can occur,
so that functionality is added by including build dependencies, rather
than setting cross-cutting compile-time #defines.  This is done by
centralizing the bus-specific logic in a "bus.cc" source file, and the
CONFIG_BRCMFMAC_{SDIO,USB,SIM} #defines only there.

This also has the effect of making the dependency graph easier to
understand, especially for tests and the new sim framework.

Changed in this commit:

* Re-implement the debugging functionality in debug.{cc,h} to minimize
  dependencies, and add it to its own minimal source_set.
* Remove Linux shim code to a separate source_set called "linuxisms".
* Remove all uses of conditional compilation, except for NDEBUG.
  generally and the use of CONFIG_BRCMFMAC_{SDIO,USB,SIM} in bus.cc in
  particular.
* Reduce the amount of cross-inclusion in header files.
* Audit all declared deps and public_deps, to minimize and regularize
  the set of each target.

TEST=paved, run brcmfmac_test, associated with GoogleGuest on vim2

Change-Id: Ie668eb1e3ced5b30c37d57f79ecd8fb50648769b