[debugger] Debugger IPC promises
Adds promises in the low-level for IPC sending. This will allow us to
build out users that uses promises from the lower-levels going outward.

A legacy Send() function is provided that wraps the new promises
version. No message-specific send calls are currently exposed. This has
the result that all IPC messages are run through promises (providing a
good indication that they're working) but no APIs have changed.

The session callback type was changed from a std::function to a
fit::function so the move-only bridge could be bound.

From this new state, we can convert the individual IPC callers to use
promises one-at-a-time.

Change ordering behavior of resolved promises. The previuos behavior
caused the debugger to malfunction as described in the comment added to
message_loop.cc. The tests were updated accordingly.

Test: There is no behavior change. All IPC messages run through this
system which means this is covered by the existing networking test.

Change-Id: Idbd17783cf34da7122b1f639748fa148c97360e2