[kernel][futex] Check state before ownership handle.
If a user attempts to wait on a futex, and declare a new owner, wait until we
have entered the futex context lock and validated the futex state before
attempting to validate the proposed new owner of the futex.  If both
the state check and the ownership check are invalid, we need to return
BAD_STATE instead of BAD_HANDLE.

This works around a possible timing edge case in how mutexes make use
of futex ownership in order to implement priority inheritance.  This
race was describe in ZX-4607, and goes like this.

1) Thread A owns mutex M and has encoded its thread handle in the M's
   futex state.
2) Thread B attempts to obtain the mutex M, but fails.  It changes the
   futex state to reflect that the mutex is owned by A _and_ is
   contested.
3) Thread B starts to call zx_futex_wait.
4) Before it gets too far into the call, thread A releases the mutex
   and exit.  Specifically, it..
4.1) Swaps UNOWNED into futex state.
4.2) Calls zx_futex_wake, because the futex was contested.  Thread B
     has not made it into the futex's wait queue yet, so it will not be
     woken.
4.3) Thread A exits, invalidating the handle value that B is
     attempting to declare to be the owner of M.
5) Thread B runs and fails the call because Thread A's handle no
   longer exists.  The error code is BAD_HANDLE
6) If the BAD_HANDLE policy is turned on for this job, we have just
   violated it.  Our thread (or perhaps even our process) is going to
   get killed.

So; write some new tests to check this behavior.  Then fix the
behavior in the kernel.

ZX-4607  #done

Change-Id: Ie239bde037a5ac8ecd2a83a754004a79b79bd58a