[ddk] Release handle ownership on Reply for devmgr FIDL transactions
The handles on the transaction are transferred on the channel write so
they need to be cleared from the fidl::Message.

This shows up when adding the test output channel, the fidl::Message
destructor attempts to clean up the channel after Reply has been called.
So this is tested via the devhost tests.

Change-Id: Icd5c2ba6cfce39a20c1f2ec40e073020c963e6a8