[graphics][compute]: Fix spinel path builder bug.
This CL fixes a small bug in the path builder code when it
allocates a new host-side "node" block, i.e. one that only
contains node-ids. All dwords in such a block are "node IDs",
i.e. tagged block IDs that can be one of:

  - The 0xFFFFFFFF value indicating an invalid node ID that
    should be ignored by the shaders. It is used to fill up
    the work-in-progress block in spn_pbi_end() with a memset()

  - A node ID with tag SPN_BLOCK_ID_TAG_PATH_NEXT, whose block
    ID will point to another "node" block. Put at the end of
    the current node block to implement a linked-list to continue
    the current path. See spn_pbi_prim_acquire_subblocks().

  - A tagged value pointing to a sub-block within  "segment"
    block holding primitive coordinates.

The variables |impl->wip.node| and |impl->wip.rem| are used
to hold a pointer into the next write position within the
current "work-in-progress" node block, as well as a count of
the remaining entries in this block. In other words, writing
a new node id is akin to:

  *impl->wip.node++ = node_id;
  impl->wip.rem --;

These values are reset everytime a new node block is
allocated, i.e.:

  - in spn_pbi_acquire_head(), which is used to allocate
    a "path head" block (same as a node block with a small
    path header struct at the start).

  - in spn_pbi_acquire_node(), which is used to allocate
    a new "node" block, which would only happen once the
    initial path head block is filled up.

The bug was in spn_pbi_acquire_node, which set |wip.rem|
to SPN_PATH_HEAD_DWORDS instead of |config.block.dwords|.

The end result was that only a few words would be filled
up in the node block. And if the block was filled with less
than SPN_PATH_HEAD_DWORDS - 1 entries, then the memset()
in spn_pbi_end() would not work correctly, leaving
undefined/uninitialized values in the rest of the block.

This could result in buggy path rendering, and even in
extremely rare case an infinite shader loop when parsing
the linked-list of path->node->node->... blocks.

BUG: None
TEST: None, since there is no way to get any rendering results
      from Spinel at the moment, and the bug would affect a
      GPU shader program.

Change-Id: I5d31c298cd910bad57785f3562050ed15d291744